Tweak particle parameter defaults
The motivation was to restore a bit of nuance to the particle colors:
They typically vary by a few palette indices, but with a large particle
'brightness' they tend look rather saturated.
The brightness was toned down to reduce the saturation, with the other particle
settings tweaked so they still have some visual pop.