Allow more precise adjustments for DRS frames per second target
FPS can now be adjusted in increments of 1 instead of 10. The maximum
FPS for the DRS target is now 250 to accomodate for 240 Hz displays,
which is becoming a viable target on high-end GPUs and low settings.

The slider can be dragged with the mouse if you need to change its
value quickly.