Fixed a typo
Typo in main.c: For some reason "*100" had been added by mistake to the end row 2677 in main.c for some reason.

main.c: Also adjusted all float values to be in for format 2.f, instead of 2.0f or just 2.0, so it is more consistent with the rest of the code.